# project-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Provider Object Storage API
  version: 0.1.0
  description: |
    Initial contract for the implemented control-plane and custom storage API.
    This draft covers the organization membership and role administration
    vertical slice together with provider setup and storage clients.
    Organization management routes require a user JWT and the
    x-organization-id header. Platform routes require a global role with the
    platform permission and do not accept an organization header as scope.
    Storage routes require an API credential in the form keyId.secret and
    enforce namespace and folder-grant boundaries.
servers:
  - url: http://localhost:4000
    description: Local API
...
paths:
...
  /v1/providers:
    get:
      tags: [Providers]
      operationId: listProviders
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      responses:
        '200':
          description: Organization-scoped providers.
          content:
            application/json:
              schema:
                type: object
                required: [providers]
                properties:
                  providers:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProviderConnection'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Providers]
      operationId: createProvider
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderCreateRequest'
      responses:
        '201':
          description: Provider created.
          content:
            application/json:
              schema:
                type: object
                required: [provider]
                properties:
                  provider:
                    $ref: '#/components/schemas/ProviderConnection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /v1/buckets:
    get:
      tags: [Buckets]
      operationId: listBucketConnections
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      responses:
        '200':
          description: Imported bucket connections.
          content:
            application/json:
              schema:
                type: object
                required: [buckets]
                properties:
                  buckets:
                    type: array
                    items:
                      $ref: '#/components/schemas/BucketConnection'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /v1/namespaces:
    get:
      tags: [Namespaces]
      operationId: listNamespaces
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      responses:
        '200':
          description: Virtual namespaces.
          content:
            application/json:
              schema:
                type: object
                required: [namespaces]
                properties:
                  namespaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Namespace'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Namespaces]
      operationId: createNamespace
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NamespaceCreateRequest'
      responses:
        '201':
          description: Namespace created.
          content:
            application/json:
              schema:
                type: object
                required: [namespace]
                properties:
                  namespace:
                    $ref: '#/components/schemas/Namespace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /v1/api-credentials:
    get:
      tags: [API credentials]
      operationId: listApiCredentials
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      responses:
        '200':
          description: Credential metadata; secrets are never returned by this operation.
          content:
            application/json:
              schema:
                type: object
                required: [credentials]
                properties:
                  credentials:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiCredential'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [API credentials]
      operationId: createApiCredential
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiCredentialCreateRequest'
      responses:
        '201':
          description: Credential metadata and one-time secret.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedApiCredential'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /v1/folder-grants:
    get:
      tags: [Folder grants]
      operationId: listFolderGrants
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
        - name: namespaceId
          in: query
          schema:
            type: string
            format: uuid
        - name: principalType
          in: query
          schema:
            type: string
            enum: [ROLE, API_CREDENTIAL]
        - name: principalId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Folder grants.
          content:
            application/json:
              schema:
                type: object
                required: [grants]
                properties:
                  grants:
                    type: array
                    items:
                      $ref: '#/components/schemas/FolderGrant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Folder grants]
      operationId: createFolderGrant
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FolderGrantCreateRequest'
      responses:
        '201':
          description: Folder grant created.
          content:
            application/json:
              schema:
                type: object
                required: [grant]
                properties:
                  grant:
                    $ref: '#/components/schemas/FolderGrant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/migrations:
    get:
      tags: [Migrations]
      operationId: listMigrations
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      responses:
        '200':
          description: Migration runs scoped to the organization.
          content:
            application/json:
              schema:
                type: object
                required: [migrations]
                properties:
                  migrations:
                    type: array
                    items: { $ref: '#/components/schemas/MigrationRun' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Migrations]
      operationId: createMigration
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MigrationCreateRequest' }
      responses:
        '201':
          description: Migration run created in CREATED state.
          content:
            application/json:
              schema:
                type: object
                required: [migration]
                properties:
                  migration: { $ref: '#/components/schemas/MigrationRun' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/TooLarge' }
...
  /v1/platform/organizations:
    get:
      tags: [Platform]
      operationId: listPlatformOrganizations
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        '200':
          description: Organizations visible to a global platform administrator.
          content:
            application/json:
              schema:
                type: object
                required: [organizations]
                properties:
                  organizations:
                    type: array
                    items: { $ref: '#/components/schemas/PlatformOrganization' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived dashboard access token.
    storageCredential:
      type: http
      scheme: bearer
      bearerFormat: keyId.secret
      description: API credential scoped to one organization and namespace.
    metricsBearer:
      type: http
      scheme: bearer
      description: Deployment-provided metrics scrape token.
  parameters:
    OrganizationIdHeader:
      name: x-organization-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthenticated:
      description: Missing, invalid, expired, revoked, or out-of-scope credential.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated principal lacks the required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource was not found in the caller's tenant scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource state or idempotency conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooLarge:
      description: The requested batch exceeds the bounded operation limit.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: string
...
    ProviderCreateRequest:
      type: object
      required: [displayName, type, endpoint, region, accessKeyId, secretAccessKey]
      additionalProperties: false
      properties:
        displayName: { type: string, minLength: 1, maxLength: 200 }
        type:
          type: string
          enum: [AWS_S3, CLOUDFLARE_R2, BACKBLAZE_B2, MINIO, GENERIC_S3]
        endpoint: { type: string, format: uri, maxLength: 2048 }
        region: { type: string, minLength: 1, maxLength: 100 }
        addressingMode:
          type: string
          enum: [AUTO, VIRTUAL_HOSTED, PATH_STYLE]
          default: AUTO
        accessKeyId: { type: string, minLength: 1, maxLength: 512, writeOnly: true }
        secretAccessKey: { type: string, minLength: 1, maxLength: 4096, writeOnly: true }
        sessionToken: { type: string, minLength: 1, maxLength: 8192, writeOnly: true }
...
    ProviderConnection:
      type: object
      required: [id, displayName, type, endpoint, region, addressingMode, status, lastHealthCheckedAt]
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        type: { type: string }
        endpoint: { type: string, format: uri }
        region: { type: string }
        addressingMode: { type: string }
        status: { type: string }
        lastHealthCheckedAt: { type: string, format: date-time, nullable: true }
...
    BucketConnection:
      type: object
      required: [id, bucketName, providerConnectionId, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        bucketName: { type: string }
        providerConnectionId: { type: string, format: uuid }
        status: { type: string }
        createdAt: { type: string, format: date-time }
    BigIntString:
      type: string
      pattern: '^\d+$'
      description: Non-negative PostgreSQL BIGINT encoded as a JSON string.
    PlacementTarget:
      type: object
      required: [id, bucketConnectionId, priorityTier, weight, thresholdPercent, enabled, configuredCapacityBytes, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        bucketConnectionId: { type: string, format: uuid }
        priorityTier: { type: integer, minimum: 0 }
        weight: { type: integer, minimum: 1 }
        thresholdPercent: { type: integer, minimum: 0, maximum: 100 }
        enabled: { type: boolean }
        configuredCapacityBytes: { $ref: '#/components/schemas/BigIntString', nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PlacementPolicy:
      type: object
      required: [id, strategy, status, targets, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        strategy: { type: string, enum: [PRIORITY_WEIGHTED] }
        status: { type: string, enum: [ACTIVE, DISABLED] }
        targets:
          type: array
          items: { $ref: '#/components/schemas/PlacementTarget' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Namespace:
      type: object
      required: [id, name, slug, status, quotaBytes, maxObjectSizeBytes, defaultTransferMode, versioningMode, placementPolicy, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        status: { type: string, enum: [ACTIVE, DISABLED] }
        quotaBytes: { $ref: '#/components/schemas/BigIntString' }
        maxObjectSizeBytes: { $ref: '#/components/schemas/BigIntString', nullable: true }
        defaultTransferMode: { type: string, enum: [DIRECT, PROXIED] }
        versioningMode: { type: string, enum: [DISABLED] }
        placementPolicy: { $ref: '#/components/schemas/PlacementPolicy', nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    NamespaceCreateRequest:
      type: object
      required: [name, slug, quotaBytes]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        slug: { type: string, pattern: '^[a-z0-9](?:[a-z0-9-]{0,98}[a-z0-9])?$' }
        quotaBytes: { $ref: '#/components/schemas/BigIntString' }
        maxObjectSizeBytes: { $ref: '#/components/schemas/BigIntString', nullable: true }
        defaultTransferMode: { type: string, enum: [DIRECT, PROXIED], default: DIRECT }
        versioningMode: { type: string, enum: [DISABLED], default: DISABLED }
...
    ApiCredential:
      type: object
      required: [id, keyId, status, createdAt, updatedAt, createdById, expiresAt, revokedAt, lastUsedAt, namespace]
      properties:
        id: { type: string, format: uuid }
        keyId: { type: string }
        status: { type: string, enum: [ACTIVE, REVOKED, EXPIRED] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        createdById: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }
        lastUsedAt: { type: string, format: date-time, nullable: true }
        namespace:
          type: object
          required: [id, name, slug]
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            slug: { type: string }
    ApiCredentialCreateRequest:
      type: object
      required: [namespaceId]
      additionalProperties: false
      properties:
        namespaceId: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time, nullable: true }
...
    CreatedApiCredential:
      type: object
      required: [credential, secret]
      properties:
        credential: { $ref: '#/components/schemas/ApiCredential' }
        secret:
          type: string
          writeOnly: true
          description: Returned only at create/rotate time; store it immediately.
    FolderGrant:
      type: object
      required: [id, namespaceId, principalType, principalId, prefix, actions, createdAt, updatedAt, createdById, expiresAt, namespace]
      properties:
        id: { type: string, format: uuid }
        namespaceId: { type: string, format: uuid }
        principalType: { type: string, enum: [ROLE, API_CREDENTIAL] }
        principalId: { type: string, format: uuid }
        prefix: { type: string, maxLength: 1024 }
        actions:
          type: array
          uniqueItems: true
          items: { type: string, enum: [list, read, write, delete] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        createdById: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time, nullable: true }
        namespace:
          type: object
          required: [id, name, slug]
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            slug: { type: string }
    FolderGrantCreateRequest:
      type: object
      required: [namespaceId, principalType, principalId, prefix, actions]
      additionalProperties: false
      properties:
        namespaceId: { type: string, format: uuid }
        principalType: { type: string, enum: [ROLE, API_CREDENTIAL] }
        principalId: { type: string, format: uuid }
        prefix: { type: string, maxLength: 1024 }
        actions:
          type: array
          minItems: 1
          maxItems: 4
          uniqueItems: true
          items: { type: string, enum: [list, read, write, delete] }
        expiresAt: { type: string, format: date-time, nullable: true }
...
    MigrationCreateRequest:
      type: object
      required: [namespaceId, sourceBucketConnectionId, targetBucketConnectionId]
      additionalProperties: false
      properties:
        namespaceId: { type: string, format: uuid }
        sourceBucketConnectionId: { type: string, format: uuid }
        targetBucketConnectionId: { type: string, format: uuid }
...
    MigrationRun:
      type: object
      required: [id, namespaceId, sourceBucketConnectionId, targetBucketConnectionId, state, totalItems, completedItems, failedItems, lastError, startedAt, completedAt, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        namespaceId: { type: string, format: uuid }
        sourceBucketConnectionId: { type: string, format: uuid }
        targetBucketConnectionId: { type: string, format: uuid }
        state: { type: string, enum: [CREATED, RUNNING, PAUSED, COMPLETED, FAILED, CANCELLED] }
        totalItems: { type: integer, minimum: 0 }
        completedItems: { type: integer, minimum: 0 }
        failedItems: { type: integer, minimum: 0 }
        lastError: { type: string, nullable: true }
        startedAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    PlatformOrganization:
      type: object
      required: [id, name, slug, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        status: { type: string, enum: [ACTIVE, SUSPENDED, CLOSED] }
        createdAt: { type: string, format: date-time }
...